authorization & authentication

+++ blog/app/Http/Controllers/BlogController.php
@@ -4,22 +4,29 @@
 namespace App\Http\Controllers;
 
 use App\Category;
+use Gate;
 use Illuminate\Http\Request;
 use App\Post;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
-
 
 /*
  * CRUD
  */
 class BlogController extends Controller
 {
+    // nurodom kad metodai pasiekiami tik prisijungusiam vartotojui, bet isvardinam isimtis
+    public function __construct(){
+        $this->middleware('auth', ['except' => ['index', 'showFull']]);
+    }
+
     public function index()
 
     {
         $posts = DB::table('posts')
             ->join('categories', 'posts.category', '=', 'categories.id')
-            ->select('posts.id', 'posts.title', 'posts.body', 'categories.category', 'categories.id')
+            ->join('users', 'posts.user_id', '=', 'users.id')
+            ->select('posts.id', 'posts.title', 'posts.body', 'categories.category_name', 'users.name') //categories.category_name
             ->paginate(5);
 
         return view('blog_theme.pages.home', compact('posts'));
@@ -32,13 +39,6 @@ class BlogController extends Controller
         return view('blog_theme.pages.add-post', compact('options'));
     }
 
-    public function edit(Post $post)
-    {
-        $options = Category::all(); // options ateina is kategoriju lenteles postai is postu
-
-        return view('blog_theme.pages.edit', compact('options', 'post'));
-    }
-
     public function store(Request $request){
         $validatedData = $request->validate([
             'title' => 'required|unique:posts|max:255',
@@ -49,7 +49,8 @@ class BlogController extends Controller
         Post::create([
             'title' => request('title'),
             'category' => request('category'),
-            'body' => request('body')
+            'body' => request('body'),
+            'user_id' => Auth::id()
         ]);
 
         return redirect('/'); // nurodom kur jam grizti i pradini psl)
@@ -60,6 +61,15 @@ class BlogController extends Controller
         return view('blog_theme/pages/posts', compact('post'));
     }
 
+    public function edit(Post $post)
+    {
+
+        if (Gate::allows('update', $post)) {
+            $options = Category::all(); // options ateina is kategoriju lenteles postai is postu
+            return view('blog_theme.pages.edit', compact('options', 'post'));
+        }
+        dd('klaida');
+    }
 
     public function storeUpdate(Request $request, Post $post)
     {
@@ -67,11 +77,15 @@ class BlogController extends Controller
         return redirect('/post/'.$post->id);
     }
 
-    public function delete(Post $post){
+    public function delete(Post $post)
+    {
 
-        $post->delete();
+        if (Gate::allows('update', $post)) {
+            $post->delete();
 
-        return redirect('/');
+            return redirect('/');
+        }
+        dd('klaida klaiduze');
     }
 
 }
